several small changes
