+++ src/app/globals.scss
@@ -78,65 +78,3 @@ body {
   color: var(--text-primary-color);
   font-family: var(--main-font-family);
 }
-
-// a {
-//   text-decoration: none;
-// }
-
-// .authContainer {
-//   width: 100vw;
-//   height: 100vh;
-//   display: flex;
-//   justify-items: center;
-//   align-items: center;
-// }
-
-// .authInnerContainer {
-//   width: 30%;
-//   height: 50%;
-//   margin: auto;
-// }
-
-// .authInnerContainer > h1 {
-//   text-align: center;
-//   padding: 1em;
-// }
-
-// .modal {
-//   display: flex;
-//   position: fixed;
-//   z-index: 1;
-//   left: 0;
-//   top: 0;
-//   width: 100%;
-//   height: 100%;
-//   overflow: auto;
-//   justify-content: center;
-//   align-items: center;
-// }
-
-// .modal-content {
-//   background-color: #333333;
-//   margin: auto;
-//   padding: 20px;
-//   border: 1px solid #888;
-//   width: 80%;
-//   max-width: 600px;
-//   box-shadow: 0 5px 15px rgba(0, 0, 0, 0.3);
-// }
-
-// .close {
-//   float: right;
-//   font-size: 20px;
-//   font-weight: bold;
-// }
-
-// .close:hover,
-// .close:focus {
-//   text-decoration: none;
-//   cursor: pointer;
-// }
-
-
-
-
