+++ src/app/login/page.js
@@ -2,9 +2,10 @@
 
 import { useState } from "react";
 import { useRouter } from "next/navigation";
-import { useUser } from "../../hooks/useUser";
-import Form from "../../components/SignupLoginForm";
+import { useUser } from "@/hooks/useUser";
+import Form from "@/components/SignupLoginForm";
 import { supabase } from "@/components/SupabaseClient";
+import Typography from "@/components/Typography";
 
 const Login = () => {
   const router = useRouter();
@@ -26,25 +27,13 @@ const Login = () => {
       const { data, error } = await supabase.auth.signInWithPassword({
         email: body.username,
         password: body.password,
-      })
+      });
       console.log(data);
       if (!error) {
         router.push("/");
       } else {
         throw new Error(error.message);
       }
-      // const url = process.env.NEXT_PUBLIC_REALTIME_SERVER_ADDRESS || "http://localhost:3030";
-      // const res = await fetch(url + "/auth/login", {
-      //   method: "POST",
-      //   credentials: 'include',
-      //   headers: { "Content-Type": "application/json" },
-      //   body: JSON.stringify(body),
-      // });
-      // if (res.status === 200) {
-      //   router.push("/");
-      // } else {
-      //   throw new Error(await res.text());
-      // }
     } catch (error) {
       console.error("An unexpected error happened occurred:", error);
       setErrorMsg(error.message);
@@ -52,11 +41,31 @@ const Login = () => {
   }
 
   return (
-    <div className="authContainer">
-      <div className="authInnerContainer">
+    <div
+      style={{
+        display: "flex",
+        flexDirection: "column",
+        justifyContent: "center",
+        alignItems: "center",
+        height: "100vh",
+      }}
+    >
+      <div
+        style={{
+          maxWidth: "1000px",
+          display: "flex",
+          flexDirection: "column",
+          alignItems: "start",
+          marginBottom: "2rem",
+        }}
+      >
+        <Typography variant="subhero">This is</Typography>
+        <Typography variant="hero">CultureHub Broadcaster</Typography>
+      </div>
+     
         <h1>Log In</h1>
         <Form isLogin errorMessage={errorMsg} onSubmit={handleSubmit} />
-      </div>
+      
     </div>
   );
 };
